+++ src/pages/index.ts
@@ -5,3 +5,4 @@
 export * from './Main';
 export * from './auth/index';
 export * from './home/index';
+export * from './setting/index';
