+++ src/pages/setting/index.ts
@@ -2,4 +2,4 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from './User';
+export * from './Setting';
